requirements.txt
anyio==4.9.0
appdirs==1.4.4
asgiref==3.8.1
click==8.2.1
connectorx==0.4.3
h11==0.16.0
htmltools==0.6.0
idna==3.10
Jinja2==3.1.6
linkify-it-py==2.0.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
narwhals==1.44.0
numpy==2.3.1
orjson==3.10.18
packaging==25.0
pandas==2.3.0
plotly==6.2.0
polars==1.31.0
prompt_toolkit==3.0.51
pyarrow==20.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
questionary==2.1.0
setuptools==80.9.0
shiny==1.4.0
six==1.17.0
sniffio==1.3.1
starlette==0.47.1
typing_extensions==4.14.0
tzdata==2025.2
uc-micro-py==1.0.3
uvicorn==0.35.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.0
